Lab 3 work and completion
Have plans to make a function which automatically reports on many different variations of preprocessing and modeling in order to help optimize models quickly. If time permits, I will continue to develop this model this semester and hopefully build out its functionality to be more interesting and layers as I move forward.